Implemented flexible training functionality:
It is now possible to train a model, save it, close the program, open it again and continue training seamlessly from where you left off, using the -train command as always, for the number of epochs specified in the config.
The model's history/plots will be saved accordingly showing the total progress.

This required some minor changes to the way models are stored, which means that continuing training of models created prior to this update is not officially supported.
If you would like to continue training an older model anyway, feel free to contact me for help. It is possible, but for several (good) reasons I decided not to include this functionality in the production code.

PLEASE NOTE: If you train a model and then continue training, the new version will replace the old version. Make sure to back-up any old versions if you would like to keep them.

Some bugs have also been fixed.